# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Recolección de Residuos",
        "description": "Documentación de la API para el sistema de gestión de rutas de recolección.",
        "contact": {
            "email": "soporte@tuempresa.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/vehiculos": {
            "get": {
                "tags": [
                    "Vehiculos"
                ],
                "summary": "Listar vehículos (filtrado opcional por perfil)",
                "description": "Devuelve una lista paginada de todos los vehículos, opcionalmente filtrada por perfil_id. Si perfil_id no se proporciona, lista todos los vehículos.",
                "operationId": "b912521b0e41dec60b1fd4d98a0e06fe",
                "parameters": [
                    {
                        "name": "perfil_id",
                        "in": "query",
                        "description": "UUID del perfil para filtrar sus vehículos",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "a1b2c3d4-e5f6-7890-abcd-1234567890ab"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de vehículos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Vehiculo"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validación fallida (si el UUID es inválido)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "El perfil_id proporcionado es inválido."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Vehiculos"
                ],
                "summary": "Crear un nuevo vehículo",
                "description": "Crea un nuevo vehículo de recolección, requiere el perfil_id en el cuerpo de la petición.",
                "operationId": "20a2f18df869f0d5e7d5ddce18285c29",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "placa",
                                    "perfil_id"
                                ],
                                "properties": {
                                    "placa": {
                                        "type": "string",
                                        "example": "ABC123"
                                    },
                                    "marca": {
                                        "type": "string",
                                        "example": "Chevrolet"
                                    },
                                    "modelo": {
                                        "type": "string",
                                        "example": "2020"
                                    },
                                    "activo": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "perfil_id": {
                                        "description": "ID del perfil al que pertenece el vehículo",
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "e6a7b8c9-d0e1-2345-fghi-678901234567"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vehículo creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Vehiculo"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validación fallida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "La placa ya existe o el perfil_id es inválido."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/vehiculos/{id}": {
            "get": {
                "tags": [
                    "Vehiculos"
                ],
                "summary": "Obtener detalles de un vehículo",
                "description": "Devuelve los detalles de un vehículo específico por su ID.",
                "operationId": "e7cea4106c3b515e1fd8f8fc20a01096",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del vehículo (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "a1b2c3d4-e5f6-7890-abcd-1234567890ab"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del vehículo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Vehiculo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vehículo no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Vehículo no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Vehiculos"
                ],
                "summary": "Actualizar un vehículo",
                "description": "Actualiza los datos de un vehículo existente usando su ID.",
                "operationId": "96084827a5107cad2b7ea8faa4eaf765",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del vehículo (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "a1b2c3d4-e5f6-7890-abcd-1234567890ab"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "placa": {
                                        "type": "string",
                                        "example": "ABC123"
                                    },
                                    "marca": {
                                        "type": "string",
                                        "example": "Chevrolet"
                                    },
                                    "modelo": {
                                        "type": "string",
                                        "example": "2020"
                                    },
                                    "activo": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "perfil_id": {
                                        "description": "Opcional: ID del perfil para reasignar el vehículo",
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "e6a7b8c9-d0e1-2345-fghi-678901234567"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vehículo actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Vehiculo"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validación fallida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "La placa ya existe."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Vehiculos"
                ],
                "summary": "Eliminar un vehículo",
                "description": "Elimina un vehículo de la base de datos por su ID.",
                "operationId": "9b11c434f19fa00c1d0b5d3fc9ca526e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del vehículo (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "a1b2c3d4-e5f6-7890-abcd-1234567890ab"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Vehículo eliminado"
                    },
                    "404": {
                        "description": "Vehículo no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Vehículo no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Vehiculo": {
                "required": [
                    "id",
                    "placa",
                    "perfil_id"
                ],
                "properties": {
                    "id": {
                        "description": "ID único del vehículo",
                        "type": "string",
                        "format": "uuid"
                    },
                    "perfil_id": {
                        "description": "ID del perfil al que pertenece el vehículo",
                        "type": "string",
                        "format": "uuid"
                    },
                    "placa": {
                        "description": "Placa del vehículo",
                        "type": "string",
                        "example": "ABC-123"
                    },
                    "marca": {
                        "description": "Marca del vehículo",
                        "type": "string",
                        "example": "Chevrolet"
                    },
                    "modelo": {
                        "description": "Año del modelo del vehículo",
                        "type": "integer",
                        "example": 2022
                    },
                    "capacidad": {
                        "description": "Capacidad de carga en toneladas",
                        "type": "number",
                        "format": "float",
                        "example": 5.5
                    },
                    "tipo_combustible": {
                        "description": "Tipo de combustible",
                        "type": "string",
                        "example": "Diésel"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}